Spread test charts across various pages
It was getting unruly throwing everything in the application.hbs
template.
