+++ test-app/app/templates/application.hbs
@@ -5,7 +5,12 @@
       <h3>Test Pages</h3>
       <ul>
         <li><LinkTo @route='index'>Index</LinkTo></li>
-        <li><LinkTo @route='foo'>Foo</LinkTo></li>
+        <li><LinkTo @route='axes'>Axes/Scales</LinkTo></li>
+        <li><LinkTo @route='lines'>Lines</LinkTo></li>
+        <li><LinkTo @route='areas'>Areas</LinkTo></li>
+        <li><LinkTo @route='arcs'>Arcs</LinkTo></li>
+        <li><LinkTo @route='points-bands'>Points/Bands</LinkTo></li>
+        <li><LinkTo @route='reactivity'>Reactivity</LinkTo></li>
       </ul>
     </div>
   </nav>
