+++ test-app/app/templates/lines.hbs
@@ -1,0 +1,152 @@
+<h2>Line</h2>
+<div class='fluid-chart'>
+  <Lineal::Fluid class='fluid-chart__plot' as |width height|>
+    <svg class='fluid-chart__svg'>
+      {{#let
+        (scale-linear range=(array 0 width))
+        (scale-linear range=(array height 0))
+        as |xScale yScale|
+      }}
+        {{#if (and xScale.isValid yScale.isValid)}}
+          <Lineal::Axis
+            @scale={{yScale}}
+            @orientation='left'
+            @tickValues={{array 20000000 100000000 200000000 280000000}}
+          />
+          <Lineal::Axis
+            @scale={{xScale}}
+            @orientation='bottom'
+            transform='translate(0,{{height}})'
+          />
+          <Lineal::Gridlines
+            @scale={{yScale}}
+            @lineValues={{array 20000000 100000000 200000000 280000000}}
+            @direction='horizontal'
+            @length={{width}}
+            stroke-dasharray='5 5'
+            opacity='0.7'
+          />
+          <Lineal::Gridlines
+            @scale={{xScale}}
+            @direction='vertical'
+            @length={{height}}
+            stroke-dasharray='5 5'
+            opacity='0.3'
+          />
+          {{#each this.population as |d|}}
+            <circle
+              cx={{xScale.compute d.year}}
+              cy={{yScale.compute d.people}}
+              r='3'
+            ></circle>
+          {{/each}}
+        {{/if}}
+        <Lineal::Line
+          @data={{this.population}}
+          @xScale={{xScale}}
+          @yScale={{yScale}}
+          @x='year'
+          @y='people'
+          @curve='natural'
+          fill='transparent'
+          stroke='black'
+          stroke-width='2'
+        />
+        <rect
+          x='0'
+          y='0'
+          width={{width}}
+          height={{height}}
+          tabindex='0'
+          fill='transparent'
+          class='interactor'
+          {{interactor-cartesian-horizontal
+            data=this.population
+            xScale=xScale
+            x='year'
+            y='people'
+            onSeek=this.updateActiveData
+          }}
+        ></rect>
+        {{#if (and xScale.isValid yScale.isValid)}}
+          {{#if this.activeDatum}}
+            <g class='interaction-overlay'>
+              <line
+                stroke='pink'
+                stroke-width='2'
+                x1={{xScale.compute this.activeDatum.year}}
+                x2={{xScale.compute this.activeDatum.year}}
+                y1='0'
+                y2={{height}}
+              ></line>
+              <circle
+                cx={{xScale.compute this.activeDatum.year}}
+                cy={{yScale.compute this.activeDatum.people}}
+                r='5'
+                fill='red'
+              ></circle>
+            </g>
+          {{/if}}
+        {{/if}}
+      {{/let}}
+    </svg>
+  </Lineal::Fluid>
+  <div>
+    {{#if this.activeDatum}}
+      <p><strong>Active Datum:</strong>
+        {{this.activeDatum.year}}</p>
+      <p><strong>Active Value:</strong>
+        {{this.activeDatum.people}}</p>
+    {{/if}}
+  </div>
+</div>
+
+<h2>Line</h2>
+<svg width='800' height='200'>
+  {{#let
+    (scale-linear range='15..785')
+    (scale-linear range='190..10')
+    as |xScale yScale|
+  }}
+    <Lineal::Area
+      @data={{this.sineFiltered}}
+      @xScale={{xScale}}
+      @yScale={{yScale}}
+      @x='x'
+      @y='y'
+      @y0={{0}}
+      fill='rgba(255,0,0,0.3)'
+      stroke-width='1'
+    />
+    <Lineal::Line
+      @data={{this.sine}}
+      @xScale={{xScale}}
+      @yScale={{yScale}}
+      @x='x'
+      @y='y'
+      fill='transparent'
+      stroke='black'
+      stroke-width='2'
+    />
+    <Lineal::Line
+      @data={{this.sineFiltered}}
+      @xScale={{xScale}}
+      @yScale={{yScale}}
+      @x='x'
+      @y='y'
+      stroke-dasharray='3 5'
+      fill='transparent'
+      stroke='black'
+      stroke-width='1'
+    />
+    {{#if (and xScale.isValid yScale.isValid)}}
+      {{#each this.sineFiltered as |d|}}
+        <circle
+          cx={{xScale.compute d.x}}
+          cy={{yScale.compute d.y}}
+          r='3'
+        ></circle>
+      {{/each}}
+    {{/if}}
+  {{/let}}
+</svg>
